% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{get_palette}
\alias{get_palette}
\title{Get Color Palette by Name}
\usage{
get_palette(palette_name)
}
\arguments{
\item{palette_name}{A string representing the name of the palette (e.g., "default", "venti").}
}
\value{
A vector of hex color codes for the requested palette.
}
\description{
This function returns the color palette based on the given name.
}
