% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{load_palettes_from_rds}
\alias{load_palettes_from_rds}
\title{Load Color Palettes from RDS File}
\usage{
load_palettes_from_rds(path)
}
\description{
This function loads a list of color palettes stored in an RDS file.
If the file is missing, an error will be thrown.
}
