% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{show_palette}
\alias{show_palette}
\title{Show Color Palette}
\usage{
show_palette(palette_name)
}
\arguments{
\item{palette_name}{A string, the name of the palette ("default", "venti", "mediterranean", "hsbc").}
}
\description{
This function displays the selected color palette.
}
