% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{get_palette}
\alias{get_palette}
\title{Get Color Palette by Name}
\usage{
get_palette(palette_name)
}
\arguments{
\item{palette_name}{A string representing the name of the palette (e.g. "default").}
}
\value{
A vector of hex color codes for the requested palette.
}
\description{
This function returns a color palette by its name.
The palette name must match one of the available palettes in the package.
}
\examples{
# Get the "hsbc" color palette
hsbc_palette <- get_palette("hsbc")
}
