% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{show_palette}
\alias{show_palette}
\title{Show Color Palette}
\usage{
show_palette(palette_name, show_names = FALSE)
}
\arguments{
\item{palette_name}{A string representing the name of the palette (e.g., "default").}

\item{show_names}{A logical value indicating whether to display the color names on the plot. Default is FALSE.}
}
\description{
This function displays a color palette as a barplot.
Optionally, color names can be shown above each color bar.
}
\examples{
# Show the "default" palette
show_palette("default")

# Show the "default" palette with color names
show_palette("default", show_names = TRUE)
}
